+++ horqrux/api.py
@@ -9,8 +9,8 @@ from jax import Array
 from jax.experimental import checkify
 
 from horqrux.composite import Observable, OpSequence
+from horqrux.differentiation.ad import ad_expectation
 from horqrux.differentiation.adjoint import adjoint_expectation as apply_adjoint
-from horqrux.differentiation.automatic_diff import ad_expectation
 from horqrux.differentiation.gpsr import finite_shots_fwd, no_shots_fwd
 from horqrux.utils import (
     DensityMatrix,
