+++ horqrux/differentiation/__init__.py
@@ -1,5 +1,5 @@
 from __future__ import annotations
 
+from .ad import ad_expectation
 from .adjoint import adjoint_expectation
-from .automatic_diff import ad_expectation
 from .gpsr import finite_shots_fwd, no_shots_fwd
