+++ horqrux/differentiation/gpsr.py
@@ -10,7 +10,7 @@ from jax.experimental import checkify
 
 from horqrux.apply import apply_gates
 from horqrux.composite import Observable
-from horqrux.differentiation.automatic_diff import _ad_expectation_single_observable
+from horqrux.differentiation.ad import _ad_expectation_single_observable
 from horqrux.primitives.primitive import Primitive
 from horqrux.utils import DensityMatrix, State, expand_operator, num_qubits
